--- src/components/StageDots/index.js
@@ -1,1 +1,0 @@
-export { default as StageDots } from './StageDots';
