+++ src/feature/account/component/StageDots/StageDots.jsx
