+++ src/feature/account/component/index.js
@@ -1,0 +1,1 @@
+export { default as StageDots } from './StageDots/StageDots';
