+++ src/pages/SnoroseVerifyPage/TermsPage/TermsPage.jsx
@@ -1,6 +1,5 @@
+import { TERMS } from '@/feature/account/constant';
 import { Button } from '@/pages/SnoroseVerifyPage';
-
-import { TERMS } from '@/constants';
 
 import styles from './TermsPage.module.css';
